requirements_optimized.txt
pandas>=2.0.0
numpy>=1.24.0
boto3>=1.26.0
pytz>=2023.3
finnhub-python>=2.4.23
python-dotenv>=1.0.0
polygon-api-client>=1.14.6
requests>=2.31.0
beautifulsoup4>=4.12.0
yfinance>=0.2.18
urllib3>=2.0.0
psutil>=5.9.0
holidays>=0.25 